app/requirements.txt
databricks-sdk>=0.81.0
fastapi>=0.104.0
uvicorn>=0.24.0
psycopg[binary]>=3.0
